Refactor conditional InvertButton render
