--- packages/lib-classifier/src/components/Classifier/components/ImageToolbar/ImageToolbarConnector.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import { withStores } from '@helpers'
-import ImageToolbar from './ImageToolbar'
-
-function storeMapper (store) {
-  const {
-    workflows: {
-      active: workflow
-    }
-  } = store
-
-  return {
-    invert: workflow?.configuration?.invert_subject
-  }
-}
-
-export default withStores(ImageToolbar, storeMapper)
