+++ packages/lib-classifier/src/components/Classifier/components/ImageToolbar/components/InvertButton/InvertButtonContainer.js
@@ -1,14 +1,35 @@
 import { withStores } from '@helpers'
+import React from 'react'
+
 import InvertButton from './InvertButton'
 
 function storeMapper(classifierStore) {
   const {
-    invertView
-  } = classifierStore.subjectViewer
+    subjectViewer: {
+      invertView
+    },
+    workflows: {
+      active: workflow
+    }
+  } = classifierStore
 
+  const disabled = !workflow?.configuration?.invert_subject
   return {
+    disabled,
     onClick: invertView
   }
 }
 
-export default withStores(InvertButton, storeMapper)
+function InvertButtonContainer ({
+  disabled = false,
+  onClick = () => console.log('invert view')
+}) {
+  if (disabled) {
+    return null
+  }
+  return (
+    <InvertButton onClick={onClick} />
+  )
+}
+
+export default withStores(InvertButtonContainer, storeMapper)
