+++ packages/lib-classifier/src/components/Classifier/components/ImageToolbar/index.js
@@ -1,1 +1,1 @@
-export { default } from './ImageToolbarConnector'
+export { default } from './ImageToolbar'
